Continued refactor

+++ Code/engineGlobals.h
@@ -2,6 +2,6 @@
 inline bool QUIT = false;
 inline float deltaTime = 0;
 
-constexpr int OBJECT_MAX = 3000;
-constexpr int LIGHT_MAX = 3000;
-constexpr int ASSET_MAX = 300;
+constexpr size_t OBJECT_MAX = 3000;
+constexpr size_t LIGHT_MAX = 3000;
+constexpr size_t ASSET_MAX = 300;
